Fix theoretical int truncation
read(2) returns a ssize_t, not an int.

NOTRY=true

Review URL: https://codereview.chromium.org/230473006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264021 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f24ffea6c08536f5a64f725063ca92c98ec51b3c

+++ components/nacl/loader/nacl_helper_linux.cc
@@ -121,7 +121,7 @@ void ChildNaClLoaderInit(const std::vector<int>& child_fds,
   // should not fork any child processes (which the seccomp
   // sandbox does) until then, because that can interfere with the
   // parent's discovery of our PID.
-  const int nread = HANDLE_EINTR(read(parent_fd, buffer, kMaxReadSize));
+  const ssize_t nread = HANDLE_EINTR(read(parent_fd, buffer, kMaxReadSize));
   const std::string switch_prefix = std::string("--") +
       switches::kProcessChannelID + std::string("=");
   const size_t len = switch_prefix.length();
@@ -129,7 +129,7 @@ void ChildNaClLoaderInit(const std::vector<int>& child_fds,
   if (nread < 0) {
     perror("read");
     LOG(ERROR) << "read returned " << nread;
-  } else if (nread > static_cast<int>(len)) {
+  } else if (static_cast<size_t>(nread) > len) {
     if (switch_prefix.compare(0, len, buffer, 0, len) == 0) {
       VLOG(1) << "NaCl loader is synchronised with Chrome zygote";
       CommandLine::ForCurrentProcess()->AppendSwitchASCII(
